feat: split and rename submit component step and update contribution flow
